[level 2] Title: 소수 찾기, Time: 14.22 ms, Memory: 80.4 MB -BaekjoonHub

+++ 프로그래머스/lv2/42839. 소수 찾기/소수 찾기.java
@@ -30,7 +30,7 @@ class Solution {
         if(number == 0 || number == 1 || number % 2 == 0)
             return false;
         
-        for(int i=3; i<=(int)Math.sqrt(number); i+=2)
+        for(int i=3; i <= Math.sqrt(number); i+=2)
             if(number % i == 0)
                 return false;
